delete old backup service and adapt breez_bridge to the new API.
