+++ lib/services/injector.dart
@@ -8,7 +8,6 @@ import 'package:breez/services/nfc.dart';
 import 'package:breez/services/permissions.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:breez/services/lightning_links.dart';
-import 'package:breez/services/backup.dart';
 
 class ServiceInjector {
   static final _singleton = new ServiceInjector._internal();
@@ -21,7 +20,6 @@ class ServiceInjector {
   DeepLinksService _deepLinksService;
   LightningLinksService _lightningLinksService;
   Device _device;
-  BackupService _backupService;
   Future<SharedPreferences> _sharedPreferences;
   Permissions _permissions;
 
@@ -53,10 +51,6 @@ class ServiceInjector {
 
   Device get device {
     return _device ??= Device();
-  }
-
-  BackupService get backupService {
-    return _backupService ??= BackupService();
   }
 
   DeepLinksService get deepLinks => _deepLinksService ??= new DeepLinksService();
